Remove v2 notifications, and remove unused notifications
